+++ src/main/java/gaussbuilder/GeometryBuilder.java
@@ -120,11 +120,11 @@ public class GeometryBuilder {
 
 				if (index == 0) {
 					if (carbons[u] == null) {
-						xu = (double) molecule.getNodesRefs()[u].getX();
-						yu = (double) molecule.getNodesRefs()[u].getY();
+						xu = molecule.getNodesRefs()[u].getX();
+						yu = molecule.getNodesRefs()[u].getY();
 					} else {
-						xu = (double) carbons[u].getX();
-						yu = (double) carbons[u].getY();
+						xu = carbons[u].getX();
+						yu = carbons[u].getY();
 					}
 
 					xv = xu + 1.2145;
@@ -134,11 +134,11 @@ public class GeometryBuilder {
 				else if (index == 1) {
 
 					if (carbons[u] == null) {
-						xu = (double) molecule.getNodesRefs()[u].getX();
-						yu = (double) molecule.getNodesRefs()[u].getY();
+						xu = molecule.getNodesRefs()[u].getX();
+						yu = molecule.getNodesRefs()[u].getY();
 					} else {
-						xu = (double) carbons[u].getX();
-						yu = (double) carbons[u].getY();
+						xu = carbons[u].getX();
+						yu = carbons[u].getY();
 					}
 
 					xv = xu;
@@ -148,11 +148,11 @@ public class GeometryBuilder {
 
 				else if (index == 2) {
 					if (carbons[u] == null) {
-						xu = (double) molecule.getNodesRefs()[u].getX();
-						yu = (double) molecule.getNodesRefs()[u].getY();
+						xu = molecule.getNodesRefs()[u].getX();
+						yu = molecule.getNodesRefs()[u].getY();
 					} else {
-						xu = (double) carbons[u].getX();
-						yu = (double) carbons[u].getY();
+						xu = carbons[u].getX();
+						yu = carbons[u].getY();
 					}
 
 					xv = xu - 1.2145;
@@ -163,11 +163,11 @@ public class GeometryBuilder {
 				else if (index == 3) {
 
 					if (carbons[u] == null) {
-						xu = (double) molecule.getNodesRefs()[u].getX();
-						yu = (double) molecule.getNodesRefs()[u].getY();
+						xu = molecule.getNodesRefs()[u].getX();
+						yu = molecule.getNodesRefs()[u].getY();
 					} else {
-						xu = (double) carbons[u].getX();
-						yu = (double) carbons[u].getY();
+						xu = carbons[u].getX();
+						yu = carbons[u].getY();
 					}
 
 					xv = xu - 1.2145;
@@ -178,11 +178,11 @@ public class GeometryBuilder {
 				else if (index == 4) {
 
 					if (carbons[u] == null) {
-						xu = (double) molecule.getNodesRefs()[u].getX();
-						yu = (double) (molecule.getNodesRefs()[u].getY());
+						xu = molecule.getNodesRefs()[u].getX();
+						yu = molecule.getNodesRefs()[u].getY();
 					} else {
-						xu = (double) carbons[u].getX();
-						yu = (double) carbons[u].getY();
+						xu = carbons[u].getX();
+						yu = carbons[u].getY();
 					}
 
 					xv = xu;
@@ -192,11 +192,11 @@ public class GeometryBuilder {
 				else if (index == 5) {
 
 					if (carbons[u] == null) {
-						xu = (double) molecule.getNodesRefs()[u].getX();
-						yu = (double) molecule.getNodesRefs()[u].getY();
+						xu = molecule.getNodesRefs()[u].getX();
+						yu = molecule.getNodesRefs()[u].getY();
 					} else {
-						xu = (double) carbons[u].getX();
-						yu = (double) carbons[u].getY();
+						xu = carbons[u].getX();
+						yu = carbons[u].getY();
 					}
 
 					xv = xu + 1.2145;
