+++ src/main/java/parsers/COMConverter2.java
@@ -318,11 +318,11 @@ public class COMConverter2 {
 
 				if (index == 0) {
 					if (carbons[u] == null) {
-						xu = (double) molecule.getNodesRefs()[u].getX();
-						yu = (double) molecule.getNodesRefs()[u].getY();
+						xu = molecule.getNodesRefs()[u].getX();
+						yu = molecule.getNodesRefs()[u].getY();
 					} else {
-						xu = (double) carbons[u].getX();
-						yu = (double) carbons[u].getY();
+						xu = carbons[u].getX();
+						yu = carbons[u].getY();
 					}
 
 					xv = xu + 1.2145;
@@ -332,11 +332,11 @@ public class COMConverter2 {
 				else if (index == 1) {
 
 					if (carbons[u] == null) {
-						xu = (double) molecule.getNodesRefs()[u].getX();
-						yu = (double) molecule.getNodesRefs()[u].getY();
+						xu = molecule.getNodesRefs()[u].getX();
+						yu = molecule.getNodesRefs()[u].getY();
 					} else {
-						xu = (double) carbons[u].getX();
-						yu = (double) carbons[u].getY();
+						xu = carbons[u].getX();
+						yu = carbons[u].getY();
 					}
 
 					xv = xu;
@@ -346,11 +346,11 @@ public class COMConverter2 {
 
 				else if (index == 2) {
 					if (carbons[u] == null) {
-						xu = (double) molecule.getNodesRefs()[u].getX();
-						yu = (double) molecule.getNodesRefs()[u].getY();
+						xu = molecule.getNodesRefs()[u].getX();
+						yu = molecule.getNodesRefs()[u].getY();
 					} else {
-						xu = (double) carbons[u].getX();
-						yu = (double) carbons[u].getY();
+						xu = carbons[u].getX();
+						yu = carbons[u].getY();
 					}
 
 					xv = xu - 1.2145;
@@ -361,11 +361,11 @@ public class COMConverter2 {
 				else if (index == 3) {
 
 					if (carbons[u] == null) {
-						xu = (double) molecule.getNodesRefs()[u].getX();
-						yu = (double) molecule.getNodesRefs()[u].getY();
+						xu = molecule.getNodesRefs()[u].getX();
+						yu = molecule.getNodesRefs()[u].getY();
 					} else {
-						xu = (double) carbons[u].getX();
-						yu = (double) carbons[u].getY();
+						xu = carbons[u].getX();
+						yu = carbons[u].getY();
 					}
 
 					xv = xu - 1.2145;
@@ -376,11 +376,11 @@ public class COMConverter2 {
 				else if (index == 4) {
 
 					if (carbons[u] == null) {
-						xu = (double) molecule.getNodesRefs()[u].getX();
-						yu = (double) (molecule.getNodesRefs()[u].getY());
+						xu = molecule.getNodesRefs()[u].getX();
+						yu = molecule.getNodesRefs()[u].getY();
 					} else {
-						xu = (double) carbons[u].getX();
-						yu = (double) carbons[u].getY();
+						xu = carbons[u].getX();
+						yu = carbons[u].getY();
 					}
 
 					xv = xu;
@@ -390,11 +390,11 @@ public class COMConverter2 {
 				else if (index == 5) {
 
 					if (carbons[u] == null) {
-						xu = (double) molecule.getNodesRefs()[u].getX();
-						yu = (double) molecule.getNodesRefs()[u].getY();
+						xu = molecule.getNodesRefs()[u].getX();
+						yu = molecule.getNodesRefs()[u].getY();
 					} else {
-						xu = (double) carbons[u].getX();
-						yu = (double) carbons[u].getY();
+						xu = carbons[u].getX();
+						yu = carbons[u].getY();
 					}
 
 					xv = xu + 1.2145;
